Create syn-ble-resize.py
An offshoot of the synthetic images generator with blemishes, this one adds an extra bit of code to resize the image to either bigger or smaller than 256. There is a problem with the original code where if I increase the size of the image, it stops the generation of the objects and throws up errors. This is a stop-gap while I solve that problem.